some cleanup.
